+++ test/generator/loader-limits.test.ts
@@ -10,7 +10,7 @@ import {
   ReasonCode,
   StringOpKind,
   ValueType,
-} from "../../generated/libbusinessid/ir/v1/rules_pb.js";
+} from "../../generated/entid/ir/v1/rules_pb.js";
 import { loadBundle } from "../../tools/generator/load.js";
 import {
   alwaysValidFormat,